拆分代码
