+++ src/TelescopeSourceGenerator/Analyzers/ExportTypeToMethod_/IExportMethodReturnTypeCollectionResult.cs
@@ -1,0 +1,9 @@
+﻿namespace dotnetCampus.Telescope.SourceGeneratorAnalyzers;
+
+/// <summary>
+/// 收集导出方法返回值类型
+/// </summary>
+/// 可以是收集到了，也可以是返回开发者定义错误代码
+interface IExportMethodReturnTypeCollectionResult
+{
+}
