+++ src/TelescopeSourceGenerator/Analyzers/ExportTypeToMethod_/TelescopeExportTypeToMethodIncrementalGenerator.cs
@@ -27,7 +27,10 @@ public class TelescopeExportTypeToMethodIncrementalGenerator : IIncrementalGener
     public void Initialize(IncrementalGeneratorInitializationContext context)
     {
 #if DEBUG
-        Debugger.Launch();
+        if (!Debugger.IsAttached)
+        {
+            Debugger.Launch();
+        }
 #endif
         // 先找到定义
         IncrementalValuesProvider<ExportTypeCollectionResult> exportMethodIncrementalValuesProvider = context.SyntaxProvider.CreateSyntaxProvider((syntaxNode, _) =>
@@ -539,164 +542,5 @@ namespace {@namespace}
         Debug.Assert(classDeclarationList != null);
         Debug.Assert(classDeclarationList!.Count > 0);
         return true;
-    }
-
-    class ExportTypeCollectionResult : IEquatable<ExportTypeCollectionResult>
-    {
-        public ExportTypeCollectionResult(IMethodSymbol methodSymbol, GeneratorSyntaxContext generatorSyntaxContext)
-        {
-            ExportPartialMethodSymbol = methodSymbol;
-            GeneratorSyntaxContext = generatorSyntaxContext;
-        }
-
-        /// <summary>
-        /// 是否包含引用的程序集和 DLL 里面的类型导出。默认只导出当前程序集
-        /// </summary>
-        public bool IncludeReference { set; get; }
-        /// <summary>
-        /// 程序集里面标记了导出的分部方法，将用来生成代码
-        /// </summary>
-        public IMethodSymbol ExportPartialMethodSymbol { get; }
-        public GeneratorSyntaxContext GeneratorSyntaxContext { get; }
-
-        public bool Equals(ExportTypeCollectionResult? other)
-        {
-            if (ReferenceEquals(null, other)) return false;
-            if (ReferenceEquals(this, other)) return true;
-
-            return GeneratorSyntaxContext.Equals(other.GeneratorSyntaxContext) && SymbolEqualityComparer.Default.Equals(ExportPartialMethodSymbol, other.ExportPartialMethodSymbol);
-        }
-
-        public override bool Equals(object? obj)
-        {
-            if (ReferenceEquals(null, obj)) return false;
-            if (ReferenceEquals(this, obj)) return true;
-            if (obj.GetType() != this.GetType()) return false;
-            return Equals((ExportTypeCollectionResult) obj);
-        }
-
-        public override int GetHashCode()
-        {
-            unchecked
-            {
-                return (SymbolEqualityComparer.Default.GetHashCode(ExportPartialMethodSymbol) * 397) ^ GeneratorSyntaxContext.GetHashCode();
-            }
-        }
-    }
-
-    /// <summary>
-    /// 收集导出方法返回值类型
-    /// </summary>
-    /// 可以是收集到了，也可以是返回开发者定义错误代码
-    interface IExportMethodReturnTypeCollectionResult
-    {
-    }
-
-    /// <summary>
-    /// 导出的方法的导出类型返回值结果
-    /// </summary>
-    class ExportMethodReturnTypeCollectionResult : IExportMethodReturnTypeCollectionResult
-    {
-        public ExportMethodReturnTypeCollectionResult(ITypeSymbol expectedClassBaseType, ITypeSymbol? expectedClassAttributeType, ExportTypeCollectionResult exportTypeCollectionResult, IExportMethodReturnTypeInfo exportMethodReturnTypeInfo)
-        {
-            ExpectedClassBaseType = expectedClassBaseType;
-            ExpectedClassAttributeType = expectedClassAttributeType;
-            ExportTypeCollectionResult = exportTypeCollectionResult;
-            ExportMethodReturnTypeInfo = exportMethodReturnTypeInfo;
-        }
-
-        /// <summary>
-        /// 期望收集的类型所继承的基础类型
-        /// </summary>
-        public ITypeSymbol ExpectedClassBaseType { get; }
-
-        /// <summary>
-        /// 期望类型标记的特性，可选
-        /// </summary>
-        public ITypeSymbol? ExpectedClassAttributeType { get; }
-
-        public ExportTypeCollectionResult ExportTypeCollectionResult { get; }
-
-        /// <summary>
-        /// 程序集里面标记了导出的分部方法，将用来生成代码
-        /// </summary>
-        public IMethodSymbol ExportPartialMethodSymbol => ExportTypeCollectionResult.ExportPartialMethodSymbol;
-
-        /// <summary>
-        /// 导出类型的返回类型信息
-        /// </summary>
-        public IExportMethodReturnTypeInfo ExportMethodReturnTypeInfo { get; }
-
-        /// <summary>
-        /// 判断传入的程序集类型满足当前的要求条件
-        /// </summary>
-        /// <param name="assemblyClassTypeSymbol"></param>
-        /// <returns></returns>
-        public bool IsMatch(INamedTypeSymbol assemblyClassTypeSymbol)
-        {
-            if (assemblyClassTypeSymbol.IsAbstract)
-            {
-                // 抽象类不能提供
-                return false;
-            }
-
-            // 先判断是否继承，再判断是否标记特性
-            if (!TypeSymbolHelper.IsInherit(assemblyClassTypeSymbol, ExpectedClassBaseType))
-            {
-                // 没有继承基类，那就是不符合了
-                return false;
-            }
-
-            if (ExpectedClassAttributeType is null)
-            {
-                // 如果没有特性要求，那就返回符合
-                return true;
-            }
-
-            foreach (var attributeData in assemblyClassTypeSymbol.GetAttributes())
-            {
-                if (SymbolEqualityComparer.Default.Equals(attributeData.AttributeClass, ExpectedClassAttributeType))
-                {
-                    return true;
-                }
-            }
-
-            return false;
-        }
-    }
-
-    class ExportMethodReturnTypeCollectionDiagnostic : IExportMethodReturnTypeCollectionResult
-    {
-    }
-
-    /// <summary>
-    /// 候选收集的结果
-    /// </summary>
-    /// 包含所有的导出结果。比如项目里面有多个导出方法，每一个导出方法就是一个 CandidateClassTypeResult 对象
-    class CandidateClassCollectionResult
-    {
-        public CandidateClassCollectionResult(IReadOnlyList<CandidateClassTypeResult> candidateClassTypeResultList)
-        {
-            CandidateClassTypeResultList = candidateClassTypeResultList;
-        }
-
-        public IReadOnlyList<CandidateClassTypeResult> CandidateClassTypeResultList { get; }
-    }
-
-    /// <summary>
-    /// 候选的导出类型结果
-    /// </summary>
-    /// 包含标记导出的方法信息，以及程序集里面导出的类型
-    class CandidateClassTypeResult
-    {
-        public CandidateClassTypeResult(ExportMethodReturnTypeCollectionResult exportMethodReturnTypeCollectionResult, IReadOnlyList<INamedTypeSymbol> assemblyClassTypeSymbolList)
-        {
-            ExportMethodReturnTypeCollectionResult = exportMethodReturnTypeCollectionResult;
-            AssemblyClassTypeSymbolList = assemblyClassTypeSymbolList;
-        }
-
-        public ExportMethodReturnTypeCollectionResult ExportMethodReturnTypeCollectionResult { get; }
-
-        public IReadOnlyList<INamedTypeSymbol> AssemblyClassTypeSymbolList { get; }
     }
 }
